Fix broken Linux build (20220920)
