+++ src/BloomExe/Utils/MiscUtils.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Diagnostics;
 using System.Drawing;
+using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Management;
@@ -9,6 +10,7 @@ using System.Security.Principal;
 using System.Text;
 using System.Windows.Forms;
 using Mono.Unix;
+using SIL.CommandLineProcessing;
 using SIL.IO;
 using SIL.PlatformUtilities;
 
@@ -459,6 +461,46 @@ namespace Bloom.Utils
 			}
 
 			Application.Idle += HandleAction;
+		}
+
+		public static TimeSpan GetMp3TimeSpan(string path)
+		{
+			try
+			{
+#if __MonoCS__
+				// ffmpeg can provide the length of the audio, but you have to strip it out of the command line output
+				// See https://stackoverflow.com/a/33115316/7442826 or https://stackoverflow.com/a/53648234/7442826
+				// The output (which is sent to stderr, not stdout) looks something like this:
+				// "size=N/A time=00:03:36.13 bitrate=N/A speed= 432x    \rsize=N/A time=00:07:13.16 bitrate=N/A speed= 433x    \rsize=N/A time=00:08:42.97 bitrate=N/A speed= 434x"
+				// When seen on the console screen interactively, it looks like a single line that is updated frequently.
+				// A short file may have only one carriage-return separated section of output, while a very long file may
+				// have more sections than this.
+				var args = String.Format("-v quiet -stats -i \"{0}\" -f null -", path);
+				var result = CommandLineRunner.Run("/usr/bin/ffmpeg", args, "", 20 * 10, new SIL.Progress.NullProgress());
+				var output = result.ExitCode == 0 ? result.StandardError : null;
+				string timeString = null;
+				if (!string.IsNullOrEmpty(output))
+				{
+					var idxTime = output.LastIndexOf("time=");
+					if (idxTime > 0)
+						timeString = output.Substring(idxTime + 5, 11);
+				}
+				return TimeSpan.Parse(timeString, CultureInfo.InvariantCulture);
+#else
+				using (var reader = new Mp3FileReader(path))
+					return reader.TotalTime;
+#endif
+			}
+			catch
+			{
+				NonFatalProblem.Report(ModalIf.All, PassiveIf.All,
+					"Bloom could not accurately determine the length of the audio file and will only make a very rough estimate.");
+				// Crude estimate. In one sample, a 61K mp3 is 7s long.
+				// So, multiply by 7 and divide by 61K to get seconds.
+				// Then, to make a TimeSpan we need ticks, which are 0.1 microseconds,
+				// hence the 10000000.
+				return new TimeSpan(new FileInfo(path).Length * 7 * 10000000 / 61000);
+			}
 		}
 	}
 }
